Extract loader into separate React component
